This is a fix for int32 that should bee uint64 because of the values it gives
modified:   src/server/game/AuctionHouse/AuctionHouseBot/AuctionHouseBot.cpp
modified:   src/server/game/AuctionHouse/AuctionHouseMgr.cpp

+++ src/server/game/AuctionHouse/AuctionHouseMgr.cpp
@@ -751,7 +751,7 @@ void AuctionEntry::SaveToDB(SQLTransaction& trans) const
     stmt->setUInt32(1, auctioneer);
     stmt->setUInt32(2, item_guidlow);
     stmt->setUInt32(3, owner);
-    stmt->setInt32 (4, int32(buyout));
+    stmt->setUInt64(4, buyout);
     stmt->setUInt64(5, uint64(expire_time));
     stmt->setUInt32(6, bidder);
     stmt->setInt32 (7, int32(bid));
